Fix(shared): Rename folder to generics
