+++ src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { DOCUMENT } from '@angular/common';
 import { Component, Inject, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { THEME_OPTIONS } from './shared/interfaces/theme.interface';
-import { UtilsService } from './shared/services/utils/utils.service';
+import { THEME_OPTIONS } from './generics/interfaces/theme.interface';
+import { UtilsService } from './generics/services/utils/utils.service';
 
 @Component({
   selector: 'app-root',
