+++ src/app/app.module.ts
@@ -5,13 +5,13 @@ import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LoaderInterceptor } from './shared/interceptors/loader/loader.interceptor';
-import { SharedModule } from './shared/shared.module';
+import { LoaderInterceptor } from './generics/interceptors/loader/loader.interceptor';
+import { GenericsModule } from './generics/shared.module';
 import { WelcomeComponent } from './welcome/welcome.component';
 
 @NgModule({
   declarations: [AppComponent, WelcomeComponent],
-  imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, HttpClientModule, SharedModule],
+  imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, HttpClientModule, GenericsModule],
   providers: [{ provide: HTTP_INTERCEPTORS, useClass: LoaderInterceptor, multi: true }],
   bootstrap: [AppComponent]
 })
