+++ src/app/auth/auth.module.ts
@@ -2,8 +2,8 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { ReactiveFormsModule } from '@angular/forms';
+import { GenericsModule } from '../generics/shared.module';
 import { MaterialModule } from '../material.module';
-import { SharedModule } from '../shared/shared.module';
 import { AuthRoutingModule } from './auth-routing.module';
 import { AuthComponent } from './auth.component';
 import { LoginComponent } from './pages/login/login.component';
@@ -11,6 +11,6 @@ import { RegisterComponent } from './pages/register/register.component';
 
 @NgModule({
   declarations: [LoginComponent, RegisterComponent, AuthComponent],
-  imports: [CommonModule, AuthRoutingModule, MaterialModule, SharedModule, ReactiveFormsModule]
+  imports: [CommonModule, AuthRoutingModule, MaterialModule, GenericsModule, ReactiveFormsModule]
 })
 export class AuthModule {}
