+++ src/app/generics/components/switcher-theme/switcher-theme.component.spec.ts
