+++ src/app/generics/data/regex.data.ts
