+++ src/app/generics/interfaces/role.interface.ts
